whisper-service/requirements.txt
fastapi
uvicorn
openai-whisper
torch
pydantic
